Sign up button programmed
The EditText and Buttons variables are created and are they are assigned to the relative controls that are in Sign Up User Interface.

The sign up button was programmed to take in the details of the user's and add them into the database created(DB)

A message is also created to notify the user of the results of the signup/insert method.