Build the Appointment Activity
Created the neccesary controls of all the buttons( Make appointment, View Appointments, <<<Return to Main) and the Textviews ( Firstname, lastname, email) and EditTexts (date and reason)

Created a new table under DB Helper and created new insert and getdata methods and created the DB class and instantiated it in the new activity.

Assigned the variables transferred from the Mainactivity into the TextViews in the AppointmentPage activties.

Created the calendar method that shows the dialog for the calendar to enter the date values

created the insert method to enter all the relevant value from the AppointmentPage activity values into the new table created called AppoitnmentDets.

Also created a getmethod and buffer method to show the details of appointments that are already made via the view data button

changed the name of layout that Mainactivity accessed to signup.xml